Beta

+++ src/main/java/com/example/autos/servicio/UsuarioServicio.java
@@ -1,10 +1,12 @@
 
 package com.example.autos.servicio;
 
+import com.example.autos.entidades.Foto;
 import com.example.autos.entidades.Usuario;
 import com.example.autos.repositorio.UsuarioRepositorio;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import javax.servlet.http.HttpSession;
 import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,9 +16,11 @@ import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
+import org.springframework.web.multipart.MultipartFile;
 
 @Service
 public class UsuarioServicio  implements UserDetailsService  {
@@ -24,9 +28,16 @@ public class UsuarioServicio  implements UserDetailsService  {
     
     @Autowired
     private UsuarioRepositorio usuarioRepo;
+    
+    @Autowired
+    private NotificacionServicio notificacionServicio;
+    private MultipartFile archivo;
+    
+@Autowired
+private FotoServicio fotoServicio;
 
     @Transactional
-    public void crearUsuario(String nombre, String apellido, String email, String clave, boolean habilitado) throws Error {
+    public void crearUsuario(MultipartFile archivo,String nombre, String apellido, String email, String clave, boolean habilitado) throws Error {
         
         validar( nombre, apellido, email, clave);
         
@@ -34,34 +45,53 @@ public class UsuarioServicio  implements UserDetailsService  {
         usuario.setNombre(nombre);
         usuario.setApellido(apellido);
         usuario.setEmail(email);
-        usuario.setClave(clave);
+        String encriptada = new BCryptPasswordEncoder().encode(clave);
+        usuario.setClave(encriptada);
         usuario.setHabilitado(habilitado);
         usuarioRepo.save(usuario);
+        notificacionServicio.enviar("Felicidades creaste tu cuenta en autos", "Autos", email);
 
+     Foto foto = fotoServicio.guardar(archivo);
+     usuario.setFoto(foto);
     }
+     
     @Transactional
-    public void modificarUsuario(Usuario usuario, String nombre, String apellido, String email, String clave, boolean habilitado) throws Error {
+    public void modificarUsuario(MultipartFile archivo,String id, String nombre, String apellido, String email, String clave, boolean habilitado) throws Error {
         
         validar( nombre, apellido, email, clave);
-        if(usuario != null){
+        Optional<Usuario> respuesta = usuarioRepo.findById(id);
+        if (respuesta.isPresent()) {
+            Usuario usuario = respuesta.get();
             usuario.setNombre(nombre);
             usuario.setApellido(apellido);
             usuario.setEmail(email);
-            usuario.setClave(clave);
+            String encriptada = new BCryptPasswordEncoder().encode(clave);
+            usuario.setClave(encriptada);
             usuario.setHabilitado(habilitado);
+            
+            String idFoto = null;
+            if(usuario.getFoto() != null){
+            idFoto = usuario.getFoto().getId();
+            }
+            Foto foto = fotoServicio.actualizar(idFoto,archivo);
+            usuario.setFoto(foto);
             usuarioRepo.save(usuario);
+            
         }else{
             throw new Error ("El usuario no puede ser nulo");
         }
+    
     }
     
-    
     @Transactional
-    public void eliminarUsuario(Usuario usuario)throws Error{
-        if(usuario != null){
-        usuarioRepo.delete(usuario);
+    public void eliminarUsuario(String id)throws Error{
+       
+        Optional<Usuario> respuesta = usuarioRepo.findById(id);
+        if (respuesta.isPresent()) {
+            Usuario usuario = respuesta.get();
+            usuarioRepo.delete(usuario);
         }else{
-            throw new Error ("El usuario no puede ser nulo");
+        throw new Error("No se encontro el id de usuario");
         }
     }
     
@@ -85,8 +115,8 @@ public class UsuarioServicio  implements UserDetailsService  {
     
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        Usuario empleado = usuarioRepo.buscarPorEmail(email);
-        if (empleado != null) {
+        Usuario usuario = usuarioRepo.buscarPorEmail(email);
+        if (usuario != null) {
 
             List<GrantedAuthority> permisos = new ArrayList<>();
 
@@ -95,13 +125,21 @@ public class UsuarioServicio  implements UserDetailsService  {
             
             ServletRequestAttributes attr =(ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
             HttpSession session = attr.getRequest().getSession(true);
-            session.setAttribute("usuario", empleado);
-            User user = new User(empleado.getEmail(), empleado.getClave(), permisos);
+            session.setAttribute("usuario", usuario);
+            User user = new User(usuario.getEmail(), usuario.getClave(), permisos);
             return user;
 
         } else {
             return null;
         }
+    }
+
+    public void modificarUsuario(String id, String nombre, String apellido, String email, String clave2, boolean chk) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
+    }
+
+    public void crearUsuario(String nombre, String apellido, String email, String clave1, boolean chk) {
+        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
 }
